Add files via upload
This is the HTML file of the python code I have done already. This html file is exported for the same python code file; in case the user does not have active access to python or the user quickly wants to check what is inside the file. So, the html version contains both code and the outputs.